+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### New Features
+* RollbackToSavePoint() in WriteBatch/WriteBatchWithIndex
+
 ### Public API Changes
 * Deprecated WriteOptions::timeout_hint_us. We no longer support write timeout. If you really need this option, talk to us and we might consider returning it.
 * Deprecated purge_redundant_kvs_while_flush option.